UCourse/certificates/views.py
import io

from rest_framework import generics, permissions, status
from rest_framework.response import Response
import os
from django.conf import settings
from django.http import HttpResponse
from django.template.loader import get_template
from xhtml2pdf import pisa
from django.contrib.staticfiles import finders

from courses.models import UserCourse
from courses.serializers import UserCourseSerializer


class GenerateCertificate(generics.GenericAPIView):

    @staticmethod
    def link_callback(uri, rel):
        """
        Convert HTML URIs to absolute system paths so xhtml2pdf can access those
        resources
        """
        result = finders.find(uri)
        if result:
            if not isinstance(result, (list, tuple)):
                result = [result]
            result = list(os.path.realpath(path) for path in result)
            path = result[0]
        else:
            sUrl = settings.STATIC_URL  # Typically /static/
            sRoot = settings.STATIC_ROOT  # Typically /home/userX/project_static/
            mUrl = settings.MEDIA_URL  # Typically /media/
            mRoot = settings.MEDIA_ROOT  # Typically /home/userX/project_static/media/

            if uri.startswith(mUrl):
                path = os.path.join(mRoot, uri.replace(mUrl, ""))
            elif uri.startswith(sUrl):
                path = os.path.join(sRoot, uri.replace(sUrl, ""))
            else:
                return uri

        # make sure that file exists
        if not os.path.isfile(path):
            raise Exception(
                'media URI must start with %s or %s' % (sUrl, mUrl)
            )
        return path

    def get(self, request, *args, **kwargs):
        template = get_template('course-certificate-template.html')

        context = {
            # 'pagesize': 'A3',
            'name': 'Hàn Phong',
            'course_name': 'Python nâng cao',
            'score': '10/10',
            'rank': 'Giỏi',
            'date': '27/08/2020'
        }
        html = template.render(context)
        result = io.BytesIO()
        pdf = pisa.pisaDocument(
            io.BytesIO(html.encode("UTF-8")), dest=result,
            link_callback=GenerateCertificate.link_callback, encoding='UTF-8'
        )
        if not pdf.err:
            response = HttpResponse(result.getvalue(), content_type='application/pdf')
            response['Content-Disposition'] = 'attachment; filename="mypdf.pdf"'
            return response
            # return HttpResponse(result.getvalue(), content_type='application/pdf')


