requirements.txt
asgiref==3.2.3
cffi==1.14.0
cryptography==2.8
dj-static==0.0.6
Django==3.0.7
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-elasticsearch-dsl==7.1.1
django-extensions==2.2.8
django-filter==2.2.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.11.0
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
Pillow==7.0.0
pycparser==2.19
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
static-ranges==0.2.0
static3==0.7.0
urllib3==1.25.8
Werkzeug==1.0.0
whitenoise==5.1.0

requests~=2.22.0